create tarball from binary rpm contents (SOFTWARE-4091)
That is, rather than extracting the tarball from the source rpm

Note we are careful here to do it in such a way that the tarball
produced will always have the same checksum if made from the same rpm.

This means making sure file timestamps, ownership, and ordering are
consistent.